pyproject.toml
[project]
name = "local-machine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Michael Booth", email = "michael@databooth.com.au" }
]
requires-python = ">=3.13"
dependencies = [
    "loguru>=0.7.3",
    "plumbum>=1.9.0",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.1",
    "watchdog>=6.0.0",
]

[project.scripts]
local-machine = "local_machine:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
